feat: block trip cancel 2 hour before departure

+++ src/main/java/ru/tututu/trains/service/ReservationService.java
@@ -3,18 +3,30 @@ package ru.tututu.trains.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.tututu.trains.entity.Reservation;
+import ru.tututu.trains.entity.TripPoint;
+import ru.tututu.trains.exceptions.NotFoundException;
 import ru.tututu.trains.model.ReservationRequestModel;
 import ru.tututu.trains.repo.ReservationRepo;
+import ru.tututu.trains.repo.TripPointRepo;
 
 import java.sql.SQLException;
+import java.sql.Time;
+import java.time.LocalTime;
+import java.time.temporal.TemporalUnit;
+import java.util.Optional;
+import java.util.concurrent.TimeUnit;
 
 @Service
 public class ReservationService {
     @Autowired
     private final ReservationRepo reservationRepo;
 
-    public ReservationService(ReservationRepo reservationRepo) {
+    @Autowired
+    private final TripPointRepo tripPointRepo;
+
+    public ReservationService(ReservationRepo reservationRepo, TripPointRepo tripPointRepo) {
         this.reservationRepo = reservationRepo;
+        this.tripPointRepo = tripPointRepo;
     }
 
     public Reservation createReservation(ReservationRequestModel reservationRequestModel, int userId) throws SQLException {
@@ -30,6 +42,15 @@ public class ReservationService {
     }
 
     public void deleteReservation(int id) throws SQLException {
+        Optional<Reservation> reservation = reservationRepo.findReservationById(id);
+        if(reservation.isEmpty())
+            throw new NotFoundException("reservation with id " + id + " not found");
+
+        TripPoint tripPoint = tripPointRepo.getTripById(reservation.get().getStartPointId()).get();
+        if(LocalTime.now().plusHours(2).isAfter(tripPoint.getDepartureTime().toLocalTime())){
+            throw new RuntimeException("could not cancel reservation");
+        }
+
         reservationRepo.deleteReservation(id);
     }
 }
